requirements.txt
Django==5.1.1
djangorestframework==3.15.2
graphene==3.4.3
graphene_django==3.2.2
rest_framework_simplejwt==0.0.2
